requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0

# Audio processing
librosa>=0.9.0
pydub>=0.25.1
soundfile>=0.10.3

# Speech recognition and diarization
SpeechRecognition>=3.8.1
pyannote.audio>=2.1.1
torch>=1.10.0
transformers>=4.18.0

# Natural language processing
nltk>=3.6.5
spacy>=3.2.0
scikit-learn>=1.0.2

# Sentiment analysis
textblob>=0.17.1
vaderSentiment>=3.3.2

# Utilities
tqdm>=4.62.3
matplotlib>=3.5.1